I misinterpreted one of the arguments to read(...). Fixed.

+++ src/java/org/catacombae/jfuse/FUSE26Operations.java
@@ -277,7 +277,7 @@ public interface FUSE26Operations extends FUSEErrorValues {
      * @param path <b>(const char*)</b> a path to the file to read from.
      * @param dest <b>(char*)</b> the destination buffer.
      * @param len <b>(size_t)</b> the number of bytes to read.
-     * @param off <b>(off_t)</b> offset in <code>dest</code> to start writing.
+     * @param off <b>(off_t)</b> offset in file to start reading.
      * @param fi <b>(struct fuse_file_info*)</b> file info.
      * @return <code>null</code> if the operation completed successfully, and an
      * <code>int</code> value otherwise, indicating what went wrong.
@@ -285,7 +285,7 @@ public interface FUSE26Operations extends FUSEErrorValues {
     public int read(byte[] path,
 		     byte[] dest,
 		     int len,
-		     int off,
+		     long off,
 		     FUSEFileInfo fi);
 
     /**
